+++ CHANGELOG.md
@@ -7,7 +7,10 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
-### [6.1.2] - 2020-10-28
+### Fixed
+- Fixed a bug where the entity API could panic.
+
+### [6.1.2, 6.1.3] - 2020-10-28
 
 ### Fixed
 - Fixed a crash in the backend and agent related to Javascript execution.